more modifications to tests
Commented out the create methods as these will likely need to be run within a beforeAll or beforeEach to ensure that records are created prior to querying them and testing the rest of the functionality.

We will also need to find a way to increment the id and remove a record within afterall to keep our tests as automated as possible.